Fixed code with cppcheck linter AEMath: intPow() added more pre-checks and fixed negative exponents AEMath: added piDouble(), another name for tau() AEUtils: removed const from return types AETimer: getWorldTime() added an explicit cast to long double to avoid number overflow

+++ AETimer/include/AETimer.hpp
@@ -145,7 +145,7 @@ public:
 	/// </summary>
 	/// <returns>double of the approximate world time the timer has counted (using it's ticks)</returns>
 	inline long double getWorldTime(void) const noexcept {
-		return this->m_ullTicks.load() * this->m_rDelay.getDelay();
+		return this->m_ullTicks.load() * (long double) this->m_rDelay.getDelay();
 	}
 
 	/// <summary>
